Josh is shite at doing unity game dev
check ResourceSpawn.cs for what i thought i'd be able to have done by now,

also, moved scripts to a new folder called Scripts,
and created a prefabs folder for the resources
