+++ Assets/Prefabs/boxPrefab.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b0e085921389a2843bc3a3dacb58caa7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
