+++ Assets/Prefabs/plankPrefab.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 556c5a44cb6af134dbc0c443b7fe233c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
