+++ Assets/Scripts/Pickup.cs
