+++ Assets/Scripts/Pickup.cs.meta
