+++ Assets/Scripts/Spawnable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd22720e624bf1f4489d57117642e314
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
